requirements.txt
psutil~=7.0.0
python-dotenv~=1.1.0
requests~=2.32.3
watchdog~=6.0.0
pyTelegramBotAPI~=4.26.0
Flask~=3.1.2
Flask-Migrate~=4.1.0
Flask-Bcrypt~=1.0.1
SQLAlchemy~=2.0.44
PyJWT~=2.10.1
alembic~=1.17.2
